requirements.txt
python-dotenv
ipykernel
synthetix==0.1.10
